refactor: cria e herda classe base para retornos da api
