+++ src/main/java/reserva_api/utils/ApiBaseReturn.java
@@ -1,0 +1,42 @@
+package reserva_api.utils;
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+public class ApiBaseReturn {
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy hh:mm:ss")
+    private LocalDateTime timestamp;
+    private String message;
+    private List<CustomError> errors;
+
+    public ApiBaseReturn() {
+        this.timestamp = LocalDateTime.now();
+    }
+
+    public LocalDateTime getTimestamp() {
+        return timestamp;
+    }
+
+    public void setTimestamp(LocalDateTime timestamp) {
+        this.timestamp = timestamp;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+
+    public void setMessage(String message) {
+        this.message = message;
+    }
+
+    public List<CustomError> getErrors() {
+        return errors;
+    }
+
+    public void setErrors(List<CustomError> errors) {
+        this.errors = errors;
+    }
+}
