+++ src/main/java/reserva_api/utils/ApiError.java
@@ -9,38 +9,16 @@ import java.util.List;
 
 
 
-public class ApiError {
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "dd-MM-yyyy hh:mm:ss")
-    private LocalDateTime timestamp;
-    private String message;
+public class ApiError extends ApiBaseReturn {
     private List<CustomError> errors;
 
     public ApiError() {
-        this.timestamp = LocalDateTime.now();
         this.errors = new ArrayList<CustomError>();
     }
 
     public ApiError(String error) {
         this();
-        this.errors = new ArrayList<CustomError>();
         this.errors.add(new CustomError(error));
-    }
-
-    public LocalDateTime getTimestamp() {
-        return timestamp;
-    }
-
-    public void setTimestamp(LocalDateTime timestamp) {
-        this.timestamp = timestamp;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
     }
 
     public List<CustomError> getErrors() {
@@ -48,10 +26,6 @@ public class ApiError {
     }
 
     public void setError(String error) {
-        if(this.errors == null) {
-            this.errors = Arrays.asList(new CustomError(error));
-        } else {
-            this.errors.add(new CustomError(error));
-        }
+        this.errors.add(new CustomError(error));
     }
 }
